BAAAAH

+++ GestionProjets/GestionProjets/Employe.cs
@@ -26,7 +26,7 @@ namespace GestionProjets
             this.matricule = r.GetString("matricule");
             this.statut = r.GetString("statut");
         }
-        public Employe(string nom, string prenom, DateTime dateNaissance, string email, string adresse, DateTime dateEmbauche, int tauxHoraire, string photo, string statut) {
+        public Employe(string nom, string prenom, DateTime dateNaissance, string email, string adresse, DateTime dateEmbauche, double tauxHoraire, string photo, string statut) {
             this.nom = nom;
             this.prenom = prenom;
             this.dateNaissance = dateNaissance;
